+++ lib/validators/kpp_format_validator.rb
@@ -1,5 +1,5 @@
-# http://vipiskaegrul.ru/slovar-terminov/kod-prichiny-postanovki-kpp.html
-class KppValidator < ValidatesRussian::Validator
+class KppFormatValidator < ValidatesRussian::Validator
+  # see format here: http://vipiskaegrul.ru/slovar-terminov/kod-prichiny-postanovki-kpp.html
   validates_using do |kpp|
     next false unless kpp.size == 9
     next false unless ValidatesRussian::REGION_NUMBERS.include?(kpp[0..1])