+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+### not released yet
+
+* InnValidator and KppValidator were renamed to InnFormat and KppFormat validators
+
 ### 0.0.3
 
-InnValidator
+* InnValidator added
 
 ### 0.0.2
 
-KppValidator
+* KppValidator added